+++ plugins/cookie_manager/lib/src/cookie_mgr.dart
@@ -28,7 +28,20 @@ class CookieManager extends Interceptor {
   final CookieJar cookieJar;
 
   /// Merge cookies into a Cookie string.
+  /// Cookies with longer paths are listed before cookies with shorter paths.
   static String getCookies(List<Cookie> cookies) {
+    // Sort cookies by path (longer path first).
+    cookies.sort((a, b) {
+      if (a.path == null && b.path == null) {
+        return 0;
+      } else if (a.path == null) {
+        return -1;
+      } else if (b.path == null) {
+        return 1;
+      } else {
+        return (b.path!.length).compareTo(a.path!.length);
+      }
+    });
     return cookies.map((cookie) => '${cookie.name}=${cookie.value}').join('; ');
   }
 
@@ -38,11 +51,11 @@ class CookieManager extends Interceptor {
       final previousCookies =
           options.headers[HttpHeaders.cookieHeader] as String?;
       final newCookies = getCookies([
-        ...cookies,
         ...?previousCookies
             ?.split(';')
             .where((e) => e.isNotEmpty)
             .map((c) => Cookie.fromSetCookieValue(c)),
+        ...cookies,
       ]);
       options.headers[HttpHeaders.cookieHeader] =
           newCookies.isNotEmpty ? newCookies : null;