+++ plugins/cookie_manager/CHANGELOG.md
@@ -3,6 +3,8 @@
 ## Unreleased
 
 - Allow `Set-Cookie` to be parsed in redirect responses.
+- Fix new cookies being replaced by old cookies with the same name.
+- Sort the cookie by path (longer path first).
 
 ## 2.1.2